requirements.txt
# puls-events-chatbot-intelligent-rag/requirements.txt

# ===============================
# LangChain (architecture RAG)
# ===============================
langchain==0.2.5
langchain-core==0.2.10
langchain-community==0.2.5

# ===============================
# RAG evaluation
# ===============================
ragas<0.1.7
# datasets  # (optionnel – uniquement si benchmarks Ragas avancés)

# ===============================
# Vector database
# ===============================
faiss-cpu==1.7.4

# ===============================
# LLM & Embeddings (Mistral)
# ===============================
mistralai==0.1.3

# ===============================
# API backend
# ===============================
fastapi==0.110.0
uvicorn==0.27.1
pydantic>=2.0

# ===============================
# Data processing
# ===============================
pandas==2.2.1
numpy==1.26.4

# ===============================
# HTTP & external APIs
# ===============================
requests>=2.31.0

# ===============================
# HuggingFace ecosystem (Ragas / datasets)
# ===============================
huggingface-hub==0.19.4
pyarrow==15.0.2

# ===============================
# Environment variables
# ===============================
python-dotenv==1.0.1

# ===============================
# Dashboard (optionnel)
# ===============================
streamlit>=1.31
